fix: SSR 추천 호출 경로 정리
