--- src/entities/experts/server/index.ts
@@ -1,3 +1,0 @@
-import 'server-only';
-
-export { getExpertRecommendationsServer } from './getExpertRecommendations.server';
